+++ packages/testkits/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @ts-linq/testkits
 
+## 7.0.27
+
+### Patch Changes
+
+- @ts-linq/provider-mssql@3.0.27
+- @ts-linq/provider-mysql@3.0.27
+- @ts-linq/provider-postgres@3.0.27
+
 ## 7.0.26
 
 ### Patch Changes